+++ hal-core/Cargo.toml
